requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
crewai==0.28.8 
crewai_tools==0.1.6 
azure-storage-blob